fix(chain): cupom minta RVM real e não poeira de 10 wei
CreateCouponAsync recebia AmountRvm como long e repassava direto ao
contrato, que minta unidades RAW (18 decimais) — 10 RVM viravam 10 wei.
A conversão ×10^18 agora fica no command (BigInteger; estoura long) e o
serviço recebe as unidades RAW prontas, espelhando o faucet do Token
(RvmConstants). O agregado continua guardando o valor legível.

+++ src/modules/coupon/Revoa.Coupon.Application/Services/ICouponChainService.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using Revoa.IntegrationContracts.UserWallets;
 
 namespace Revoa.Coupon.Application.Services;
@@ -11,10 +12,11 @@ public interface ICouponChainService
     Task EnsureFaucetCouponAdminRoleAsync(CancellationToken ct = default);
 
     // Assinado pela FAUCET (COUPON_ADMIN_ROLE): createCoupon(code, amount, maxUses, expiryUnix).
+    // `amount` já em unidades RAW do RVM (18 decimais) — 10 RVM = 10×10^18 (estoura long).
     // Retorna (codeHash lido do evento CouponCreated, txHash).
     Task<(string codeHash, string txHash)> CreateCouponAsync(
         string code,
-        long amount,
+        BigInteger amount,
         int maxUses,
         long expiryUnix,
         CancellationToken ct = default);
